replace files_in_subtree with collect_files
