+++ cli/fmt.rs
@@ -9,6 +9,7 @@
 
 use crate::colors;
 use crate::diff::diff;
+use crate::fs::{collect_files, is_supported_ext};
 use crate::text_encoding;
 use deno_core::error::generic_error;
 use deno_core::error::AnyError;
@@ -23,7 +24,6 @@ use std::path::Path;
 use std::path::PathBuf;
 use std::sync::atomic::{AtomicUsize, Ordering};
 use std::sync::{Arc, Mutex};
-use walkdir::WalkDir;
 
 const BOM_CHAR: char = '\u{FEFF}';
 
@@ -40,7 +40,7 @@ pub async fn format(
     return format_stdin(check);
   }
   // collect the files that are to be formatted
-  let target_files = collect_files(args, exclude)?;
+  let target_files = collect_files(args, exclude, is_supported_ext)?;
   let config = get_config();
   if check {
     check_source_files(config, target_files).await
@@ -199,56 +199,6 @@ fn files_str(len: usize) -> &'static str {
   }
 }
 
-fn is_supported(path: &Path) -> bool {
-  let lowercase_ext = path
-    .extension()
-    .and_then(|e| e.to_str())
-    .map(|e| e.to_lowercase());
-  if let Some(ext) = lowercase_ext {
-    ext == "ts" || ext == "tsx" || ext == "js" || ext == "jsx" || ext == "mjs"
-  } else {
-    false
-  }
-}
-
-pub fn collect_files(
-  files: Vec<PathBuf>,
-  ignore: Vec<PathBuf>,
-) -> Result<Vec<PathBuf>, std::io::Error> {
-  let mut target_files: Vec<PathBuf> = vec![];
-
-  // retain only the paths which exist and ignore the rest
-  let canonicalized_ignore: Vec<PathBuf> = ignore
-    .into_iter()
-    .filter_map(|i| i.canonicalize().ok())
-    .collect();
-
-  let files = if files.is_empty() {
-    vec![std::env::current_dir()?]
-  } else {
-    files
-  };
-
-  for file in files {
-    for entry in WalkDir::new(file)
-      .into_iter()
-      .filter_entry(|e| {
-        e.path().canonicalize().map_or(false, |c| {
-          !canonicalized_ignore.iter().any(|i| c.starts_with(i))
-        })
-      })
-      .filter_map(|e| match e {
-        Ok(e) if !e.file_type().is_dir() && is_supported(e.path()) => Some(e),
-        _ => None,
-      })
-    {
-      target_files.push(entry.into_path().canonicalize()?)
-    }
-  }
-
-  Ok(target_files)
-}
-
 fn get_config() -> dprint::configuration::Configuration {
   use dprint::configuration::*;
   ConfigurationBuilder::new().deno().build()
@@ -329,76 +279,5 @@ where
     Err(e)
   } else {
     Ok(())
-  }
-}
-
-#[cfg(test)]
-mod tests {
-  use super::*;
-  use std::fs;
-  use tempfile::TempDir;
-
-  #[test]
-  fn test_is_supported() {
-    assert!(!is_supported(Path::new("tests/subdir/redirects")));
-    assert!(!is_supported(Path::new("README.md")));
-    assert!(is_supported(Path::new("lib/typescript.d.ts")));
-    assert!(is_supported(Path::new("cli/tests/001_hello.js")));
-    assert!(is_supported(Path::new("cli/tests/002_hello.ts")));
-    assert!(is_supported(Path::new("foo.jsx")));
-    assert!(is_supported(Path::new("foo.tsx")));
-    assert!(is_supported(Path::new("foo.TS")));
-    assert!(is_supported(Path::new("foo.TSX")));
-    assert!(is_supported(Path::new("foo.JS")));
-    assert!(is_supported(Path::new("foo.JSX")));
-    assert!(is_supported(Path::new("foo.mjs")));
-    assert!(!is_supported(Path::new("foo.mjsx")));
-  }
-
-  #[test]
-  fn test_collect_files() {
-    fn create_files(dir_path: &PathBuf, files: &[&str]) {
-      fs::create_dir(dir_path).expect("Failed to create directory");
-      for f in files {
-        let path = dir_path.join(f);
-        fs::write(path, "").expect("Failed to create file");
-      }
-    }
-
-    // dir.ts
-    // ├── a.ts
-    // ├── b.js
-    // ├── child
-    // │   ├── e.mjs
-    // │   ├── f.mjsx
-    // │   ├── .foo.TS
-    // │   └── README.md
-    // ├── c.tsx
-    // ├── d.jsx
-    // └── ignore
-    //     ├── g.d.ts
-    //     └── .gitignore
-
-    let t = TempDir::new().expect("tempdir fail");
-
-    let root_dir_path = t.path().join("dir.ts");
-    let root_dir_files = ["a.ts", "b.js", "c.tsx", "d.jsx"];
-    create_files(&root_dir_path, &root_dir_files);
-
-    let child_dir_path = root_dir_path.join("child");
-    let child_dir_files = ["e.mjs", "f.mjsx", ".foo.TS", "README.md"];
-    create_files(&child_dir_path, &child_dir_files);
-
-    let ignore_dir_path = root_dir_path.join("ignore");
-    let ignore_dir_files = ["g.d.ts", ".gitignore"];
-    create_files(&ignore_dir_path, &ignore_dir_files);
-
-    let result =
-      collect_files(vec![root_dir_path], vec![ignore_dir_path]).unwrap();
-    let expected = [".foo.TS", "e.mjs", "d.jsx", "c.tsx", "b.js", "a.ts"];
-    for e in expected.iter() {
-      assert!(result.iter().find(|r| r.ends_with(e)).is_some());
-    }
-    assert_eq!(result.len(), expected.len());
   }
 }
